// client/src/data/Ranking.tsx
export type Ranking = {
  azeritePowers: any[];
  characterID: number;
  characterName: string;
  class: string;
  difficulty: number;
  duration: number;
  encounterID: number;
  encounterName: string;
  estimated: boolean;
  fightID: number;
  gear: any[];
  ilvlKeyOrPatch: number;
  outOf: number;
  percentile: number;
  rank: number;
  reportID: string;
  server: string;
  spec: string;
  startTime: number;
  talents: any[];
  total: number;
};